# xadminapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "adminapi",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin_role/create_admin_role": {
            "post": {
                "tags": [
                    "后台角色"
                ],
                "summary": "新增角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.CreateAdminRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/admin_role/delete_admin_role": {
            "delete": {
                "tags": [
                    "后台角色"
                ],
                "summary": "删除角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.DeleteAdminRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/admin_role/get_admin_role": {
            "get": {
                "tags": [
                    "后台角色"
                ],
                "summary": "获取角色列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "角色名",
                        "name": "role_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "运营商",
                        "name": "seller_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.XAdminRole"
                            }
                        }
                    }
                }
            }
        },
        "/admin_role/update_admin_role": {
            "patch": {
                "tags": [
                    "后台角色"
                ],
                "summary": "更新角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.UpdateAdminRoleReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
        "/seller/create_seller": {
            "post": {
                "tags": [
                    "运营商"
                ],
                "summary": "新增运营商",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.CreateXSellerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/seller/delete_seller": {
            "delete": {
                "tags": [
                    "运营商"
                ],
                "summary": "删除运营商",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.DeleteXSellerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/seller/get_seller": {
            "get": {
                "tags": [
                    "运营商"
                ],
                "summary": "获取运营商列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "运营商",
                        "name": "seller_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "运营商名称",
                        "name": "seller_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.XSeller"
                            }
                        }
                    }
                }
            }
        },
        "/seller/update_seller": {
            "patch": {
                "tags": [
                    "运营商"
                ],
                "summary": "更新运营商",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "VerifyCode",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_admin.UpdateXSellerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.XAdminRole": {
            "type": "object",
            "properties": {
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "自增Id",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "parent": {
                    "description": "上级角色",
                    "type": "string"
                },
                "role_data": {
                    "description": "权限数据",
                    "type": "string"
                },
                "role_name": {
                    "description": "角色名",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "state": {
                    "description": "状态 1开启,2关闭",
                    "type": "integer"
                },
                "update_at": {
                    "description": "删除时间",
                    "type": "string"
                }
            }
        },
...
        "model.XSeller": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "自增Id",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "seller_name": {
                    "description": "运营商名称",
                    "type": "string"
                },
                "state": {
                    "description": "状态 1开启 2关闭",
                    "type": "integer"
                }
            }
        },
...
        "service_admin.CreateAdminRoleReq": {
            "type": "object",
            "required": [
                "parent",
                "role_data",
                "role_name",
                "seller_id",
                "state"
            ],
            "properties": {
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "parent": {
                    "description": "上级角色",
                    "type": "string"
                },
                "role_data": {
                    "description": "权限数据",
                    "type": "string"
                },
                "role_name": {
                    "description": "角色",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "state": {
                    "description": "状态 1开启,2关闭",
                    "type": "integer"
                }
            }
        },
...
        "service_admin.CreateXSellerReq": {
            "type": "object",
            "required": [
                "seller_id",
                "seller_name",
                "state"
            ],
            "properties": {
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "seller_name": {
                    "description": "运营商名称",
                    "type": "string"
                },
                "state": {
                    "description": "状态 1开启,2关闭",
                    "type": "integer"
                }
            }
        },
        "service_admin.DeleteAdminRoleReq": {
            "type": "object",
            "required": [
                "role_name",
                "seller_id"
            ],
            "properties": {
                "role_name": {
                    "description": "角色",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                }
            }
        },
...
        "service_admin.DeleteXSellerReq": {
            "type": "object",
            "required": [
                "seller_id"
            ],
            "properties": {
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                }
            }
        },
...
        "service_admin.UpdateAdminRoleReq": {
            "type": "object",
            "required": [
                "role_name",
                "seller_id"
            ],
            "properties": {
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "parent": {
                    "description": "上级角色",
                    "type": "string"
                },
                "role_data": {
                    "description": "权限数据",
                    "type": "string"
                },
                "role_name": {
                    "description": "角色",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "state": {
                    "description": "状态 1开启,2关闭",
                    "type": "integer"
                }
            }
        },
...
        "service_admin.UpdateXSellerReq": {
            "type": "object",
            "required": [
                "seller_id"
            ],
            "properties": {
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "seller_id": {
                    "description": "运营商",
                    "type": "integer"
                },
                "seller_name": {
                    "description": "运营商名称",
                    "type": "string"
                },
                "state": {
                    "description": "状态 1开启,2关闭",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}